Updated data structures and routing logic
